feat: Added Countdown Timer
The added countdown timer that starts when the app is displayed to encourage the user to answer the question that is showing. The countdown will pause when the user loses focus on the app and then start up again when the user opens the app up again. This was done by adding the onrecieve modifer to resign when the app is exited